% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_read_out_of_bound}
\alias{rcpp_read_out_of_bound}
\title{read out of bound}
\usage{
rcpp_read_out_of_bound(sizeofarray)
}
\arguments{
\item{sizeofarray}{gives size of array}
}
\value{
index of array out of bound
}
\description{
read out of bound
}
