% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_use_after_deallocate}
\alias{rcpp_use_after_deallocate}
\title{use after deallocate}
\usage{
rcpp_use_after_deallocate(size)
}
\arguments{
\item{size}{of array}
}
\value{
index of array after deleting it
}
\description{
use after deallocate
}
