% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_use_after_free}
\alias{rcpp_use_after_free}
\title{use after free}
\usage{
rcpp_use_after_free(size_free)
}
\arguments{
\item{size_free}{free size of the array}
}
\value{
index of array after freeing it
}
\description{
use after free
}
