% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_write_index_outofbound}
\alias{rcpp_write_index_outofbound}
\title{write index outofbound}
\usage{
rcpp_write_index_outofbound(boundvalue)
}
\arguments{
\item{boundvalue}{value stored in index}
}
\value{
stack_array index greater than size
}
\description{
write index outofbound
}
