% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_zero_sized_array}
\alias{rcpp_zero_sized_array}
\title{zero sized array}
\usage{
rcpp_zero_sized_array(vectorvalue)
}
\arguments{
\item{vectorvalue}{value at index zero}
}
\value{
ptr value at zero index
}
\description{
zero sized array
}
