\name{testUBSAN-package}
\alias{testUBSAN-package}
\alias{testUBSAN}
\docType{package}
\title{\packageTitle{testUBSAN}}
\description{\packageDescription{testUBSAN}}
\section{Package Content}{\packageIndices{testUBSAN}}
\author{\packageAuthor{testUBSAN}}
\section{Maintainer}{\packageMaintainer{testUBSAN}}
\keyword{package}
